+++ Lib/cluster/src/vq.h
@@ -2,9 +2,9 @@
 #define _VQ_H
 
 int double_tvq(double* obs, double* code_book, int Nobs, int Ncodes, 
-        int Nfeatures, int* codes, double* lowest_dist);
+        int Nfeatures, long long* codes, double* lowest_dist);
 
 int float_tvq(float* obs, float* code_book, int Nobs, int Ncodes, 
-        int Nfeatures, int* codes, float* lowest_dist);
+        int Nfeatures, long long* codes, float* lowest_dist);
 
 #endif
